Moved functions related to handling nested dicts to utils.py.
